Move method definition parsing's tests into its own file.
